pyproject.toml
[project]
name = "py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mysql-connector-python>=9.3.0",
    "pygame>=2.6.1",
]
